Make sure CommandComplete is received before moving to next command
Fixes a bug in host.executor() where we moved to next command after any
Command Complete Event was received. If we first receive CC for
unexpected event, we moved back to waiting for next command to execute
instead of waiting for CC for current command. This caused the pending
command to never complete and thus the goroutine waiting for command
to be executed would never advance.