Add support for file output
With -output-file command line option, output is written to file. If -json is
given as well, the JSON output is not indented. This will allow easier
parsing of the JSON data.

As a special case -output-file - will print the data to stdout, but not
indent the JSON data.

Related to #2