Add timeout for HCI command execution
We need to have some timeout for HCI command execution as the Command Complete
event might get lost and never arrive. If this case we'll be stuck waiting
for the Command Complete Event and will not send any new commands.

With this, we'll fail the command if we do not receive Command Complete
within 30 seconds.

Related to #4